+++ AWS-IoT-Arduino-Yun-Library/aws_iot_version.h
@@ -17,7 +17,7 @@
 #define version_h
 
 #define VERSION_MAJOR 2
-#define VERSION_MINOR 0
+#define VERSION_MINOR 1
 #define VERSION_PATCH 0
 #define VERSION_TAG ""