+++ AWS-IoT-Python-Runtime/lib/protocol/paho/client.py
@@ -51,6 +51,7 @@ sys.path.append("./lib/")
 sys.path.append("../lib/")
 import securedWebsocket.securedWebsocketCore as wssCore
 import util.progressiveBackoffCore as backoffCore
+import util.offlinePublishQueue as offlinePublishQueue
 
 VERSION_MAJOR=1
 VERSION_MINOR=0
@@ -135,6 +136,10 @@ MQTT_ERR_AUTH = 11
 MQTT_ERR_ACL_DENIED = 12
 MQTT_ERR_UNKNOWN = 13
 MQTT_ERR_ERRNO = 14
+
+# MessageQueueing DropBehavior
+MSG_QUEUEING_DROP_OLDEST = 0
+MSG_QUEUEING_DROP_NEWEST = 1
 
 if sys.version_info[0] < 3:
     sockpair_data = "0"
@@ -462,9 +467,9 @@ class Client(object):
         self._ping_t = 0
         self._last_mid = 0
         self._state = mqtt_cs_new
+        self._max_inflight_messages = 20
         self._out_messages = []
         self._in_messages = []
-        self._max_inflight_messages = 20
         self._inflight_messages = 0
         self._will = False
         self._will_topic = ""
@@ -502,9 +507,6 @@ class Client(object):
         self._tls_version = tls_version
         self._tls_insecure = False
         self._useSecuredWebsocket = useSecuredWebsocket  # Do we enable secured websocket
-        self._subscribePool = dict()
-        self._resubDone = True
-        self._resubDoneMutex = threading.Lock()
         self._backoffCore = backoffCore.progressiveBackoffCore()  # Init the backoffCore using default configuration
 
     def __del__(self):
@@ -520,7 +522,6 @@ class Client(object):
         * Raise ValueError if input params are malformed
         """
         self._backoffCore.configTime(srcBaseReconnectTimeSecond, srcMaximumReconnectTimeSecond, srcMinimumConnectTimeSecond)
-
 
     def reinitialise(self, client_id="", clean_session=True, userdata=None):
         if self._ssl:
@@ -945,7 +946,7 @@ class Client(object):
                 elif qos == 2:
                     message.state = mqtt_ms_wait_for_pubrec
                 self._out_message_mutex.release()
-
+                    
                 rc = self._send_publish(message.mid, message.topic, message.payload, message.qos, message.retain, message.dup)
 
                 # remove from inflight messages so it will be send after a connection is made
@@ -953,7 +954,7 @@ class Client(object):
                     with self._out_message_mutex:
                         self._inflight_messages -= 1
                         message.state = mqtt_ms_publish
-
+                        
                 return (rc, local_mid)
             else:
                 message.state = mqtt_ms_queued;
@@ -1033,14 +1034,12 @@ class Client(object):
             if topic is None or len(topic) == 0:
                 raise ValueError('Invalid topic.')
             topic_qos_list = [(topic.encode('utf-8'), qos)]
-            self._subscribePool[topic] = qos
         elif isinstance(topic, tuple):
             if topic[1]<0 or topic[1]>2:
                 raise ValueError('Invalid QoS level.')
             if topic[0] is None or len(topic[0]) == 0 or not isinstance(topic[0], str):
                 raise ValueError('Invalid topic.')
             topic_qos_list = [(topic[0].encode('utf-8'), topic[1])]
-            self._subscribePool[topic[0]] = topic[1]
         elif isinstance(topic, list):
             topic_qos_list = []
             for t in topic:
@@ -1049,7 +1048,6 @@ class Client(object):
                 if t[0] is None or len(t[0]) == 0 or not isinstance(t[0], str):
                     raise ValueError('Invalid topic.')
                 topic_qos_list.append((t[0].encode('utf-8'), t[1]))
-                self._subscribePool[t[0]] = t[1]
 
         if topic_qos_list is None:
             raise ValueError("No topic specified, or incorrect topic type.")
@@ -1082,14 +1080,12 @@ class Client(object):
             if len(topic) == 0:
                 raise ValueError('Invalid topic.')
             topic_list = [topic.encode('utf-8')]
-            self._subscribePool.pop(topic, None)
         elif isinstance(topic, list):
             topic_list = []
             for t in topic:
                 if len(t) == 0 or not isinstance(t, str):
                     raise ValueError('Invalid topic.')
                 topic_list.append(t.encode('utf-8'))
-                self._subscribePool.pop(t, None)
 
         if topic_list is None:
             raise ValueError("No topic specified, or incorrect topic type.")
@@ -1300,7 +1296,7 @@ class Client(object):
                     if not retry_first_connection:
                         raise
                     self._easy_log(MQTT_LOG_DEBUG, "Connection failed, retrying")
-                    time.sleep(1)
+                    self._backoffCore.backOff()
             else:
                 break
 
@@ -1771,15 +1767,7 @@ class Client(object):
             else:
                 raise TypeError('payload must be a string, unicode or a bytearray.')
 
-        # Check if the publish needs to wait because of an on-going re-subscription
-        rc = MQTT_ERR_SUCCESS
-        self._resubDoneMutex.acquire()
-        if self._resubDone:
-            rc = self._packet_queue(PUBLISH, packet, mid, qos)
-        # If we are still waiting for SUBACK for re-subscription, do NOT add to packet queue
-        self._resubDoneMutex.release()
-
-        return rc
+        return self._packet_queue(PUBLISH, packet, mid, qos)
 
     def _send_pubrec(self, mid):
         self._easy_log(MQTT_LOG_DEBUG, "Sending PUBREC (Mid: "+str(mid)+")")
@@ -1859,7 +1847,7 @@ class Client(object):
     def _send_disconnect(self):
         return self._send_simple_command(DISCONNECT)
 
-    def _send_subscribe(self, dup, topics, force=False):
+    def _send_subscribe(self, dup, topics):
         remaining_length = 2
         for t in topics:
             remaining_length = remaining_length + 2+len(t[0])+1
@@ -1873,11 +1861,7 @@ class Client(object):
         for t in topics:
             self._pack_str16(packet, t[0])
             packet.extend(struct.pack("B", t[1]))
-        if force:
-            self._resubDoneMutex.acquire()
-            self._resubDone = False  # Should wait for SUBACK for re-subscription
-            self._resubDoneMutex.release()
-        return (self._packet_queue(command, packet, local_mid, 1, force), local_mid)
+        return (self._packet_queue(command, packet, local_mid, 1), local_mid)
 
     def _send_unsubscribe(self, dup, topics):
         remaining_length = 2
@@ -1958,7 +1942,7 @@ class Client(object):
         self._messages_reconnect_reset_out()
         self._messages_reconnect_reset_in()
 
-    def _packet_queue(self, command, packet, mid, qos, force=False):
+    def _packet_queue(self, command, packet, mid, qos):
         mpkt = dict(
             command = command,
             mid = mid,
@@ -1968,12 +1952,7 @@ class Client(object):
             packet = packet)
 
         self._out_packet_mutex.acquire()
-        if force:  # Forced to be the first
-            self._current_out_packet = None  # Reset the current out packet
-            self._out_packet = []
-            self._out_packet.insert(0, mpkt)  # Put it as the first one in the packet_queue
-        else:
-            self._out_packet.append(mpkt)
+        self._out_packet.append(mpkt)
         if self._current_out_packet_mutex.acquire(False):
             if self._current_out_packet is None and len(self._out_packet) > 0:
                 self._current_out_packet = self._out_packet.pop(0)
@@ -2078,18 +2057,6 @@ class Client(object):
         # Start counting for stable connection
         self._backoffCore.startStableConnectionTimer()
 
-        #  Resubscribe
-        if result == 0:
-            subList = list()
-            for key in self._subscribePool.keys():
-                currentTuple = (key, self._subscribePool.get(key))
-                subList.append(currentTuple)
-            if subList:  # If we need to re-subscribe...
-                self._send_subscribe(False, subList, True)  # Forced to subscribe first
-                rc = self.loop_write()
-                if rc != 0:
-                    return rc  # Never block loop I/O when there is an error
-
         if result == 0:
             rc = 0
             self._out_message_mutex.acquire()
@@ -2145,9 +2112,6 @@ class Client(object):
             return MQTT_ERR_PROTOCOL
 
     def _handle_suback(self):
-        self._resubDoneMutex.acquire()
-        self._resubDone = True
-        self._resubDoneMutex.release()
         self._easy_log(MQTT_LOG_DEBUG, "Received SUBACK")
         pack_format = "!H" + str(len(self._in_packet['packet'])-2) + 's'
         (mid, packet) = struct.unpack(pack_format, self._in_packet['packet'])